% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemblystat.R
\name{assemblystat}
\alias{assemblystat}
\title{Generate assembly statistics from input genomes assemblies (FASTA)}
\usage{
assemblystat(
  fastalocation,
  outputType = "png",
  width = 700,
  height = 700,
  res = 150,
  geom_point_size = 3,
  n50colorfill = "#0072B2",
  n90colorfill = "#D55E00"
)
}
\arguments{
\item{fastalocation}{Location of the folder with only fasta files. The files
must have extensions with either \code{.fasta} or \code{.fa} format}

\item{outputType}{Output result file in pdf/tiff/jpeg/png format (default="tiff")}

\item{width}{width of the device (default=2000)}

\item{height}{height of the device (default=2000)}

\item{res}{Nominal resolution in ppi (default=250)}

\item{geom_point_size}{point size (default=3)}

\item{n50colorfill}{Hex color code of N50 plot (default= "#0072B2")}

\item{n90colorfill}{Hex color code of N90 plot (default= "#D55E00")}
}
\description{
\code{assemblystat()} generates basic assembly stats which includes
N50 size, N90 size and Genome assembly size. This function also
generates Assembly Size vs N50 plot and Assembly Size vs N50 plot
}
\examples{

assemblystat()
assemblystat("/home/user/CPgeneProfiler/testData/fasta", outputType="png", width = 700, height = 700, res = 150, geom_point_size=3, n50colorfill = "#0072B2", n90colorfill = "#D55E00")
}
