% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocarriage.R
\name{cocarriage}
\alias{cocarriage}
\title{Find cocarriage of CP genes in the input genomes}
\usage{
cocarriage(cpgcov = 100, cpgpident = 100)
}
\arguments{
\item{cpgcov}{CP gene coverage cutoff (default=100)}

\item{cpgpident}{CP gene percentage identity with genomic sequence (default=100)}
}
\description{
\code{cocarriage()} finds if 2 or more CP genes exists in same contig
or multiple contigs. This function should be used after running
\code{filt_blast()}.
}
\examples{

cocarriage()
cocarriage(cpgcov=100,cpgpident=100)
}
