% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpblast.R
\name{cpblast}
\alias{cpblast}
\title{Generate NCBI BLAST Results by aligning input genome assemblies against Carbapenamase (CP) gene database}
\usage{
cpblast(fastalocation, dblocation, num_threads = "16", evalue = "1e-6")
}
\arguments{
\item{fastalocation}{Location of the folder with only fasta files. The files
must end with either \code{.fasta} or \code{.fa} format}

\item{dblocation}{Location of the \code{NCBI_BARRGD_CPG_DB.fasta} (CP gene database)}

\item{num_threads}{Number of threads to run the blast (default=16)}

\item{evalue}{Cutoff e-value for blast hit (default="1e-6")}
}
\description{
Prior to running this command, the \code{NCBI_BARRGD_CPG_DB.fasta} file is expected to be
downloaded and saved in the respective location on the local system. \code{cpblast()} helps
to generate the location of the CP genes in the genomes.
}
\examples{

cpblast("/home/user/CPgeneProfiler/testData/fasta","/home/user/CPgeneProfiler/testData/db")
cpblast(fastalocation = "/home/user/CPgeneProfiler/testData/fasta",dblocation = "/home/user/CPgeneProfiler/testData/db",num_threads = 8,evalue = "1e-6")
}
