% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPgeneProfiler.R
\name{CPgeneProfiler}
\alias{CPgeneProfiler}
\title{Generate a profile of carbapenamase genes from the genome assemblies}
\usage{
CPgeneProfiler(fastafiles_location)
}
\arguments{
\item{fastafiles_location}{Need a folder location with all the fasta files to check for CP genes}
}
\description{
CPgene-profiler checks for a list of CarbaPenamase (CP) genes from a list of
genome assemblies provided in fasta file format. It reports the profile of all
the CP genes available in the genome assemblies in the format of simple heatmap.
Apart from this, it also reports the presence of co-carriage of CP genes within
an assembly. Other assembly statistics such as N50, N90, Assembly Size from the
assembly are calculated and plots of length distribution of CP gene contigs from
the list of assemblies are reported.
}
