% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpprofile.R
\name{cpprofile}
\alias{cpprofile}
\title{Carbapenamase Gene Profile}
\usage{
cpprofile(
  outputType = "png",
  width = 2000,
  height = 2000,
  res = 250,
  xlab = "Carbapenamase Genes",
  ylab = "Assembly",
  title = "Carbapenamase Gene Profile Heatmap",
  titlesize = 15,
  labelsize = 12,
  colorcode_low = "#143D59",
  colorcode_high = "#F4B41A",
  cpgcov = 100,
  cpgpident = 100
)
}
\arguments{
\item{outputType}{Output result file in pdf/tiff/jpeg/png format (default="tiff")}

\item{width}{width of the device (default=2000)}

\item{height}{height of the device (default=2000)}

\item{res}{Nominal resolution in ppi (default=250)}

\item{xlab}{Label on x-axis (default="Carbapenamase Genes")}

\item{ylab}{Label on y-axis (default="Assembly")}

\item{title}{Title of the plot (default="Carbapenamase Gene Profile Heatmap")}

\item{titlesize}{Title size (default=15)}

\item{labelsize}{Label size (default=12)}

\item{colorcode_low}{Hex color code for low values (default="#143D59")}

\item{colorcode_high}{Hex color code for high values (default="#F4B41A")}

\item{cpgcov}{CP gene coverage cutoff (default=100)}

\item{cpgpident}{CP gene percentage identity with genomic sequence (default=100)}
}
\description{
\code{cpprofile()} creates a heatmap of carbapenamase gene profile from the input genome assemblies
}
\examples{

cpprofile()
cpprofile(outputType="png", width = 2000, height = 2000, res = 250, xlab="Carbapenamase Genes", ylab="Assembly", title="Carbapenamase Gene Profile Heatmap", titlesize=15, labelsize=12,colorcode_low = "#143D59", colorcode_high = "#F4B41A", cpgcov=100, cpgpident=100)
}
