% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_summary.R
\name{db_summary}
\alias{db_summary}
\title{CPgeneProfiler Database Summary}
\usage{
db_summary()
}
\description{
Display the details of Database, which includes Database Name, Database Version,
Total sequences in Databases, Date on which database was created, Database Reference location
from where sequences are downloaded
}
\examples{
db_summary()
}
