% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files2move.R
\name{files2move}
\alias{files2move}
\title{Move the output files for summarizing CPgeneProfiler Results}
\usage{
files2move(files.to.move, outdir, subdir)
}
\arguments{
\item{files.to.move}{files to move to specific directory}

\item{outdir}{Main output folder directory (default: CPgeneProfiler_Output)}

\item{subdir}{Subfolder in the \code{CPgeneProfiler_Output} directory for which files has to move}
}
\description{
Used for summarizing all the results from CPgeneProfiler output folder
}
\examples{

}
