% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filt_blast.R
\name{filt_blast}
\alias{filt_blast}
\title{Filter BLAST results}
\usage{
filt_blast(cpgcov = 100, cpgpident = 100)
}
\arguments{
\item{cpgcov}{CP gene coverage cutoff (default=100)}

\item{cpgpident}{CP gene percentage identity with genomic sequence (default=100)}
}
\description{
\code{filt_blast} filters the output results obtained from \code{cpblast()}.
The filter is based on CP gene coverage and Percentage Identity
}
\examples{

filt_blast()
filt_blast(cpgcov=100,cpgpident=100)
}
