% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpeg2pdf.R
\name{jpeg2pdf}
\alias{jpeg2pdf}
\title{Converts jpeg images to pdf for summary report}
\usage{
jpeg2pdf(outdir = "CPgeneProfiler_Output", report = "Summary")
}
\description{
Summarize all the results from CPgeneProfiler output
}
