% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpeg_merge.R
\name{jpeg_merge}
\alias{jpeg_merge}
\title{Merge jpeg images to pdf for summary report}
\usage{
jpeg_merge(pdfFile, jpegFiles)
}
\description{
Function used for summarizing all the results from CPgeneProfiler output
}
