% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_conlen.R
\name{plot_conlen}
\alias{plot_conlen}
\title{Plot CP gene contig length distribution}
\usage{
plot_conlen(
  outputType = "png",
  width = 700,
  height = 700,
  res = 150,
  xlab = "Contig Length",
  ylab = "Number of Contigs",
  title = " Contig Length Distribution",
  element_text_angle = 90,
  unit = "KB",
  breaks = 15,
  colorfill = "#F99245",
  cpgcov = 100,
  cpgpident = 100
)
}
\arguments{
\item{outputType}{Output result file in pdf/tiff/jpeg/png format (default="tiff")}

\item{width}{width of the device (default=700)}

\item{height}{height of the device (default=700)}

\item{res}{Nominal resolution in ppi (default=150)}

\item{xlab}{Label on x-axis (default="Contig Length")}

\item{ylab}{Label on y-axis (default="Number of Contigs")}

\item{title}{Title of the plot (default=" Contig Length Distribution")}

\item{element_text_angle}{Angle of element text(default=90),}

\item{unit}{x-axis units KB/MB (default="KB")}

\item{breaks}{Number of breaks (default=15)}

\item{colorfill}{bar color (default= "#F99245")}

\item{cpgcov}{CP gene coverage cutoff (default=100)}

\item{cpgpident}{CP gene percentage identity with genomic sequence (default=100)}
}
\description{
\code{plot_conlen} generates length distribution for all the CP gene
contigs across the input genome assemblies.
}
\examples{

plot_conlen()
plot_conlen(outputType="tiff", width = 700, height = 700, res = 150, xlab="Contig Length", ylab="Number of Contigs", title=" Contig Length Distribution",element_text_angle=90,unit="KB", breaks=15, colorfill = "#F99245",cpgcov=100, cpgpident=100)
}
