% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiff2pdf.R
\name{tiff2pdf}
\alias{tiff2pdf}
\title{Converts tiff images to pdf for summary report}
\usage{
tiff2pdf(outdir = "CPgeneProfiler_Output", report = "Summary")
}
\description{
Summarize all the results from CPgeneProfiler output
}
