% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiff_merge.R
\name{tiff_merge}
\alias{tiff_merge}
\title{Merge tiff images to pdf for summary report}
\usage{
tiff_merge(pdfFile, tiffFiles)
}
\description{
Function used for summarizing all the results from CPgeneProfiler output
}
