% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsetR_plot.R
\name{upsetR_plot}
\alias{upsetR_plot}
\title{UpsetR - Set intersection plot visualization}
\usage{
upsetR_plot(
  outputType = "png",
  width = 2000,
  height = 2000,
  res = 250,
  xlab = "Carbapenamase Gene Set Size",
  ylab = "Number of genome assemblies",
  cpgcov = 100,
  cpgpident = 100,
  order.by = "degree",
  nsets = 40,
  number.angles = 0,
  point.size = 1.5,
  line.size = 1,
  sets.bar.color = "red"
)
}
\arguments{
\item{outputType}{Output result file in pdf/tiff/jpeg/png format (default="tiff")}

\item{width}{width of the device (default=2000)}

\item{height}{height of the device (default=2000)}

\item{res}{Nominal resolution in ppi (default=250)}

\item{xlab}{Label on x-axis (default="Carbapenamase Genes")}

\item{ylab}{Label on y-axis (default="Assembly")}

\item{cpgcov}{CP gene coverage cutoff (default=100)}

\item{cpgpident}{CP gene percentage identity with genomic sequence (default=100)}

\item{order.by}{How the intersections in the matrix should be ordered by (default="degree")}

\item{nsets}{Number of sets to look at (default=40)}

\item{number.angles}{The angle of the numbers atop the intersection size bars (default = 0)}

\item{point.size}{Size of points in matrix plot (default=1.5)}

\item{line.size}{Width of lines in matrix plot (default = 1)}

\item{sets.bar.color}{Color of set size bar plot (default= "red")}
}
\description{
\code{upsetR_plot()} generates set intersection plot of CP genes across all the input
genome assemblies.
}
\examples{

upsetR_plot()
upsetR_plot(outputType="png", width = 2000, height = 2000, res = 250, xlab="Carbapenamase Gene Set Size", ylab="Number of genome assemblies",cpgcov=100, cpgpident=100, order.by = "degree",nsets = 40, number.angles = 0,point.size = 1.5, line.size = 1,sets.bar.color = "red")
}
